+++ src/app/Components/Navbar.js
@@ -4,7 +4,7 @@ import Image from "next/image";
 export default function Navbar() {
   return (
     <header className="bg-[#dfaf90] flex w-full items-center h-screen max-h-14 justify-between">
-      {/*/Pages/Home?LandingPage*/}
+      
       <Link href="/">
         <div className="flex mx-4 justify-between items-center ">
           <Image src="/phLogo.png" height="30" width="40" alt="Yum Yummers" />
