--- src/app/Pages/Verify/page.js
@@ -1,17 +1,0 @@
-"client side"
-
-export default function Verification() {
-return(
-<div className="h-screen w-fill">
-<div className="bg-gray-500 h-30">
-</div>
-
-
-    
-    
-    
-
-
-
-</div>
-)}
